Initial Design
Here is the first version of the cat feeder design I had. The motor was not designed by me and it is a place holder as I find a suitable motor to be used. The wmv file should show off all of the parts. This design was preliminary to get a understanding on what I wanted to make.